refactor: update file structure
chore: update scripts
