+++ content/pages/blog/index.mdx
@@ -1,0 +1,6 @@
+---
+title: Blog
+template: 'page'
+---
+
+Lorem ipsum Blog
