+++ content/pages/contact/index.mdx
@@ -1,0 +1,6 @@
+---
+title: Contact
+template: 'page'
+---
+
+Lorem ipsum Contact
