+++ src/components/page/index.jsx
@@ -2,19 +2,14 @@ import * as React from 'react'
 import { graphql, Link } from 'gatsby'
 import { MDXProvider } from '@mdx-js/react'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
-import Layout from './layout'
-import SEO from './seo'
+import Layout from '../layout'
+import SEO from '../seo'
 
 const shortcodes = { Link }
-const Post = ({ data: { mdx } }) => {
+const Page = ({ data: { mdx } }) => {
   return (
     <Layout>
-      <SEO
-        title={mdx.frontmatter.title}
-        description={mdx.excerpt}
-        pathname={mdx.slug}
-        canonicalUrl={mdx.frontmatter.canonicalUrl}
-      />
+      <SEO title={mdx.frontmatter.title} description={mdx.excerpt} />
       <h1>{mdx.frontmatter.title}</h1>
       <MDXProvider components={shortcodes}>
         <MDXRenderer>{mdx.body}</MDXRenderer>
@@ -22,21 +17,18 @@ const Post = ({ data: { mdx } }) => {
     </Layout>
   )
 }
+
 export const pageQuery = graphql`
-  query PostQuery($id: String) {
+  query PageQuery($id: String) {
     mdx(id: { eq: $id }) {
       id
       body
       excerpt
-      fields {
-        slug
-      }
       frontmatter {
         title
-        canonicalUrl
       }
     }
   }
 `
 
-export default Post
+export default Page
