+++ src/components/post/index.jsx
@@ -2,14 +2,19 @@ import * as React from 'react'
 import { graphql, Link } from 'gatsby'
 import { MDXProvider } from '@mdx-js/react'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
-import Layout from './layout'
-import SEO from './seo'
+import Layout from '../layout'
+import SEO from '../seo'
 
 const shortcodes = { Link }
-const Page = ({ data: { mdx } }) => {
+const Post = ({ data: { mdx } }) => {
   return (
     <Layout>
-      <SEO title={mdx.frontmatter.title} description={mdx.excerpt} />
+      <SEO
+        title={mdx.frontmatter.title}
+        description={mdx.excerpt}
+        pathname={mdx.slug}
+        canonicalUrl={mdx.frontmatter.canonicalUrl}
+      />
       <h1>{mdx.frontmatter.title}</h1>
       <MDXProvider components={shortcodes}>
         <MDXRenderer>{mdx.body}</MDXRenderer>
@@ -17,18 +22,21 @@ const Page = ({ data: { mdx } }) => {
     </Layout>
   )
 }
-
 export const pageQuery = graphql`
-  query PageQuery($id: String) {
+  query PostQuery($id: String) {
     mdx(id: { eq: $id }) {
       id
       body
       excerpt
+      fields {
+        slug
+      }
       frontmatter {
         title
+        canonicalUrl
       }
     }
   }
 `
 
-export default Page
+export default Post
